+++ src/Util.hs
@@ -2,6 +2,9 @@
 
 module Util where
 
+import Data.Char (isSpace)
+import Data.List (dropWhileEnd)
+
 (<&>) :: Functor f => f a -> (a -> b) -> f b
 (<&>) = flip (<$>)
 
@@ -13,3 +16,8 @@ updateLast f (a:as) = loop a as
   where
   loop b []       = [f b]
   loop b (c : cs) = b : loop c cs
+
+-- | Delete whitespace from both ends.
+
+trim :: String -> String
+trim = dropWhileEnd isSpace . dropWhile isSpace