+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Revision history for agda2lagda
 
+## (unreleased)
+
+* Render doubly-underlined (`===`) paragraphs as `\heading`,
+  dash-underlined (`---`) paragraphs as `\subheading`.
+* Tested with GHC 8.10.4 and 9.0.1.
+
 ## 0.2020.11.1
 
 * First version. Released Halloween 2020.
 * Converts agda/hs files into lagda/lhs LaTeX literate files,
   turning line comments into text and block comments into
   LaTeX comments.
+* Tested with GHC 8.0.2 - 8.10.3.